+++ basinkit/__init__.py
@@ -31,7 +31,7 @@ from .exceptions import (
     OutletSnapError,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.3.0"
 
 __all__ = [
     "Basin",